+++ src/index.ts
@@ -23,6 +23,7 @@ export * from './core/expectErr';
 
 // Conversions
 export * from './core/fromPromise';
+export * from './core/fromSerialized';
 export * from './core/fromNullable';
 export * from './core/try';
 export * from './core/toPromise';